+++ ov_msckf/cmake/ROS2.cmake
@@ -34,11 +34,11 @@ list(APPEND thirdparty_libraries
         ${Boost_LIBRARIES}
         ${CERES_LIBRARIES}
         ${OpenCV_LIBRARIES}
-)
+        )
 list(APPEND ament_libraries
         rclcpp
         tf2_ros
-	tf2_geometry_msgs
+        tf2_geometry_msgs
         std_msgs
         geometry_msgs
         sensor_msgs
@@ -47,7 +47,7 @@ list(APPEND ament_libraries
         image_transport
         ov_core
         ov_init
-)
+        )
 
 ##################################################
 # Make the shared library
@@ -60,12 +60,13 @@ list(APPEND LIBRARY_SOURCES
         src/state/StateHelper.cpp
         src/state/Propagator.cpp
         src/core/VioManager.cpp
+        src/core/VioManagerHelper.cpp
         src/update/UpdaterHelper.cpp
         src/update/UpdaterMSCKF.cpp
         src/update/UpdaterSLAM.cpp
         src/update/UpdaterZeroVelocity.cpp
-)
-list(APPEND LIBRARY_SOURCES src/ros/ROS2Visualizer.cpp)
+        )
+list(APPEND LIBRARY_SOURCES src/ros/ROS2Visualizer.cpp src/ros/ROSVisualizerHelper.cpp)
 file(GLOB_RECURSE LIBRARY_HEADERS "src/*.h")
 add_library(ov_msckf_lib SHARED ${LIBRARY_SOURCES} ${LIBRARY_HEADERS})
 ament_target_dependencies(ov_msckf_lib ${ament_libraries})
@@ -75,11 +76,11 @@ install(TARGETS ov_msckf_lib
         LIBRARY DESTINATION lib
         RUNTIME DESTINATION bin
         PUBLIC_HEADER DESTINATION include
-)
+        )
 install(DIRECTORY src/
         DESTINATION include
         FILES_MATCHING PATTERN "*.h" PATTERN "*.hpp"
-)
+        )
 ament_export_include_directories(include)
 ament_export_libraries(ov_msckf_lib)
