+++ ov_msckf/src/core/VioManager.cpp
@@ -21,6 +21,28 @@
 
 #include "VioManager.h"
 
+#include "feat/Feature.h"
+#include "feat/FeatureDatabase.h"
+#include "feat/FeatureInitializer.h"
+#include "track/TrackAruco.h"
+#include "track/TrackDescriptor.h"
+#include "track/TrackKLT.h"
+#include "track/TrackSIM.h"
+#include "types/Landmark.h"
+#include "types/LandmarkRepresentation.h"
+#include "utils/opencv_lambda_body.h"
+#include "utils/print.h"
+#include "utils/sensor_data.h"
+
+#include "init/InertialInitializer.h"
+
+#include "state/Propagator.h"
+#include "state/State.h"
+#include "state/StateHelper.h"
+#include "update/UpdaterMSCKF.h"
+#include "update/UpdaterSLAM.h"
+#include "update/UpdaterZeroVelocity.h"
+
 using namespace ov_core;
 using namespace ov_type;
 using namespace ov_msckf;
@@ -129,6 +151,27 @@ VioManager::VioManager(VioManagerOptions &params_) : thread_init_running(false),
 
   // Feature initializer for active tracks
   active_tracks_initializer = std::make_shared<FeatureInitializer>(params.featinit_options);
+}
+
+void VioManager::feed_measurement_imu(const ov_core::ImuData &message) {
+
+  // Get the oldest camera timestamp that we can remove IMU measurements before
+  // Then push back to our propagator and pass the IMU time we can delete up to
+  double oldest_time = trackFEATS->get_feature_database()->get_oldest_timestamp();
+  if (oldest_time != -1) {
+    oldest_time += params.calib_camimu_dt;
+  }
+  propagator->feed_imu(message, oldest_time);
+
+  // Push back to our initializer
+  if (!is_initialized_vio) {
+    initializer->feed_imu(message, oldest_time);
+  }
+
+  // Push back to the zero velocity updater if we have it
+  if (is_initialized_vio && updaterZUPT != nullptr) {
+    updaterZUPT->feed_imu(message, oldest_time);
+  }
 }
 
 void VioManager::feed_measurement_simulation(double timestamp, const std::vector<int> &camids,
@@ -609,289 +652,4 @@ void VioManager::do_feature_propagate_update(const ov_core::CameraData &message)
                  calib->quat()(3), calib->pos()(0), calib->pos()(1), calib->pos()(2));
     }
   }
-}
-
-bool VioManager::try_to_initialize(const ov_core::CameraData &message) {
-
-  // Directly return if the initialization thread is running
-  // Note that we lock on the queue since we could have finished an update
-  // And are using this queue to propagate the state forward. We should wait in this case
-  if (thread_init_running) {
-    std::lock_guard<std::mutex> lck(camera_queue_init_mtx);
-    camera_queue_init.push_back(message.timestamp);
-    return false;
-  }
-
-  // If the thread was a success, then return success!
-  if (thread_init_success) {
-    return true;
-  }
-
-  // Run the initialization in a second thread so it can go as slow as it desires
-  thread_init_running = true;
-  std::thread thread([&] {
-    // Returns from our initializer
-    double timestamp;
-    Eigen::MatrixXd covariance;
-    std::vector<std::shared_ptr<ov_type::Type>> order;
-    auto init_rT1 = boost::posix_time::microsec_clock::local_time();
-
-    // Try to initialize the system
-    // We will wait for a jerk if we do not have the zero velocity update enabled
-    // Otherwise we can initialize right away as the zero velocity will handle the stationary case
-    bool wait_for_jerk = (updaterZUPT == nullptr);
-    bool success = initializer->initialize(timestamp, covariance, order, state->_imu, wait_for_jerk);
-
-    // If we have initialized successfully we will set the covariance and state elements as needed
-    // TODO: set the clones and SLAM features here so we can start updating right away...
-    if (success) {
-
-      // Set our covariance (state should already be set in the initializer)
-      StateHelper::set_initial_covariance(state, covariance, order);
-
-      // Set the state time
-      state->_timestamp = timestamp;
-      startup_time = timestamp;
-
-      // Cleanup any features older than the initialization time
-      // Also increase the number of features to the desired amount during estimation
-      trackFEATS->get_feature_database()->cleanup_measurements(state->_timestamp);
-      trackFEATS->set_num_features(params.num_pts);
-      if (trackARUCO != nullptr) {
-        trackARUCO->get_feature_database()->cleanup_measurements(state->_timestamp);
-      }
-
-      // If we are moving then don't do zero velocity update4
-      if (state->_imu->vel().norm() > params.zupt_max_velocity) {
-        has_moved_since_zupt = true;
-      }
-
-      // Else we are good to go, print out our stats
-      auto init_rT2 = boost::posix_time::microsec_clock::local_time();
-      PRINT_INFO(GREEN "[init]: successful initialization in %.4f seconds\n" RESET, (init_rT2 - init_rT1).total_microseconds() * 1e-6);
-      PRINT_INFO(GREEN "[init]: orientation = %.4f, %.4f, %.4f, %.4f\n" RESET, state->_imu->quat()(0), state->_imu->quat()(1),
-                 state->_imu->quat()(2), state->_imu->quat()(3));
-      PRINT_INFO(GREEN "[init]: bias gyro = %.4f, %.4f, %.4f\n" RESET, state->_imu->bias_g()(0), state->_imu->bias_g()(1),
-                 state->_imu->bias_g()(2));
-      PRINT_INFO(GREEN "[init]: velocity = %.4f, %.4f, %.4f\n" RESET, state->_imu->vel()(0), state->_imu->vel()(1), state->_imu->vel()(2));
-      PRINT_INFO(GREEN "[init]: bias accel = %.4f, %.4f, %.4f\n" RESET, state->_imu->bias_a()(0), state->_imu->bias_a()(1),
-                 state->_imu->bias_a()(2));
-      PRINT_INFO(GREEN "[init]: position = %.4f, %.4f, %.4f\n" RESET, state->_imu->pos()(0), state->_imu->pos()(1), state->_imu->pos()(2));
-
-      // Remove any camera times that are order then the initialized time
-      // This can happen if the initialization has taken a while to perform
-      std::lock_guard<std::mutex> lck(camera_queue_init_mtx);
-      std::vector<double> camera_timestamps_to_init;
-      for (size_t i = 0; i < camera_queue_init.size(); i++) {
-        if (camera_queue_init.at(i) > timestamp) {
-          camera_timestamps_to_init.push_back(camera_queue_init.at(i));
-        }
-      }
-
-      // Now we have initialized we will propagate the state to the current timestep
-      // In general this should be ok as long as the initialization didn't take too long to perform
-      // Propagating over multiple seconds will become an issue if the initial biases are bad
-      size_t clone_rate = (size_t)((double)camera_timestamps_to_init.size() / (double)params.state_options.max_clone_size) + 1;
-      for (size_t i = 0; i < camera_timestamps_to_init.size(); i += clone_rate) {
-        propagator->propagate_and_clone(state, camera_timestamps_to_init.at(i));
-        StateHelper::marginalize_old_clone(state);
-      }
-      PRINT_DEBUG(YELLOW "[init]: moved the state forward %.2f seconds\n" RESET, state->_timestamp - timestamp);
-      thread_init_success = true;
-      camera_queue_init.clear();
-
-    } else {
-      auto init_rT2 = boost::posix_time::microsec_clock::local_time();
-      PRINT_DEBUG(YELLOW "[init]: failed initialization in %.4f seconds\n" RESET, (init_rT2 - init_rT1).total_microseconds() * 1e-6);
-      thread_init_success = false;
-      std::lock_guard<std::mutex> lck(camera_queue_init_mtx);
-      camera_queue_init.clear();
-    }
-
-    // Finally, mark that the thread has finished running
-    thread_init_running = false;
-  });
-
-  // If we are single threaded, then run single threaded
-  // Otherwise detach this thread so it runs in the background!
-  if (!params.use_multi_threading) {
-    thread.join();
-  } else {
-    thread.detach();
-  }
-  return false;
-}
-
-void VioManager::retriangulate_active_tracks(const ov_core::CameraData &message) {
-
-  // Start timing
-  boost::posix_time::ptime retri_rT1, retri_rT2, retri_rT3, retri_rT4, retri_rT5;
-  retri_rT1 = boost::posix_time::microsec_clock::local_time();
-
-  // Clear old active track data
-  active_tracks_time = state->_timestamp;
-  active_image = message.images.at(0).clone();
-  active_tracks_posinG.clear();
-  active_tracks_uvd.clear();
-
-  // Get all features which are tracked in the current frame
-  // NOTE: This database should have all features from all trackers already in it
-  // NOTE: it also has the complete history so we shouldn't see jumps from deleting measurements
-  std::vector<std::shared_ptr<Feature>> active_features = trackDATABASE->features_containing_older(state->_timestamp);
-
-  // 0. Get all timestamps our clones are at (and thus valid measurement times)
-  std::vector<double> clonetimes;
-  for (const auto &clone_imu : state->_clones_IMU) {
-    clonetimes.emplace_back(clone_imu.first);
-  }
-
-  // 1. Clean all feature measurements and make sure they all have valid clone times
-  //    Also remove any that we are unable to triangulate (due to not having enough measurements)
-  auto it0 = active_features.begin();
-  while (it0 != active_features.end()) {
-
-    // Skip if it is a SLAM feature since it already is already going to be added
-    if (state->_features_SLAM.find((*it0)->featid) != state->_features_SLAM.end()) {
-      it0 = active_features.erase(it0);
-      continue;
-    }
-
-    // Clean the feature
-    (*it0)->clean_old_measurements(clonetimes);
-
-    // Count how many measurements
-    int ct_meas = 0;
-    for (const auto &pair : (*it0)->timestamps) {
-      ct_meas += (*it0)->timestamps[pair.first].size();
-    }
-
-    // Remove if we don't have enough and am not a SLAM feature which doesn't need triangulation
-    if (ct_meas < (int)std::max(4.0, std::floor(state->_options.max_clone_size * 2.0 / 5.0))) {
-      it0 = active_features.erase(it0);
-    } else {
-      it0++;
-    }
-  }
-  retri_rT2 = boost::posix_time::microsec_clock::local_time();
-
-  // Return if no features
-  if (active_features.empty() && state->_features_SLAM.empty())
-    return;
-
-  // 2. Create vector of cloned *CAMERA* poses at each of our clone timesteps
-  std::unordered_map<size_t, std::unordered_map<double, FeatureInitializer::ClonePose>> clones_cam;
-  for (const auto &clone_calib : state->_calib_IMUtoCAM) {
-
-    // For this camera, create the vector of camera poses
-    std::unordered_map<double, FeatureInitializer::ClonePose> clones_cami;
-    for (const auto &clone_imu : state->_clones_IMU) {
-
-      // Get current camera pose
-      Eigen::Matrix3d R_GtoCi = clone_calib.second->Rot() * clone_imu.second->Rot();
-      Eigen::Vector3d p_CioinG = clone_imu.second->pos() - R_GtoCi.transpose() * clone_calib.second->pos();
-
-      // Append to our map
-      clones_cami.insert({clone_imu.first, FeatureInitializer::ClonePose(R_GtoCi, p_CioinG)});
-    }
-
-    // Append to our map
-    clones_cam.insert({clone_calib.first, clones_cami});
-  }
-  retri_rT3 = boost::posix_time::microsec_clock::local_time();
-
-  // 3. Try to triangulate all features that have measurements
-  auto it1 = active_features.begin();
-  while (it1 != active_features.end()) {
-
-    // Triangulate the feature and remove if it fails
-    bool success_tri = true;
-    if (active_tracks_initializer->config().triangulate_1d) {
-      success_tri = active_tracks_initializer->single_triangulation_1d(it1->get(), clones_cam);
-    } else {
-      success_tri = active_tracks_initializer->single_triangulation(it1->get(), clones_cam);
-    }
-
-    // Remove the feature if not a success
-    if (!success_tri) {
-      it1 = active_features.erase(it1);
-      continue;
-    }
-    it1++;
-  }
-  retri_rT4 = boost::posix_time::microsec_clock::local_time();
-
-  // Return if no features
-  if (active_features.empty() && state->_features_SLAM.empty())
-    return;
-
-  // Points which we have in the global frame
-  for (const auto &feat : active_features) {
-    active_tracks_posinG[feat->featid] = feat->p_FinG;
-  }
-  for (const auto &feat : state->_features_SLAM) {
-    Eigen::Vector3d p_FinG = feat.second->get_xyz(false);
-    if (LandmarkRepresentation::is_relative_representation(feat.second->_feat_representation)) {
-      // Assert that we have an anchor pose for this feature
-      assert(feat.second->_anchor_cam_id != -1);
-      // Get calibration for our anchor camera
-      Eigen::Matrix3d R_ItoC = state->_calib_IMUtoCAM.at(feat.second->_anchor_cam_id)->Rot();
-      Eigen::Vector3d p_IinC = state->_calib_IMUtoCAM.at(feat.second->_anchor_cam_id)->pos();
-      // Anchor pose orientation and position
-      Eigen::Matrix3d R_GtoI = state->_clones_IMU.at(feat.second->_anchor_clone_timestamp)->Rot();
-      Eigen::Vector3d p_IinG = state->_clones_IMU.at(feat.second->_anchor_clone_timestamp)->pos();
-      // Feature in the global frame
-      p_FinG = R_GtoI.transpose() * R_ItoC.transpose() * (feat.second->get_xyz(false) - p_IinC) + p_IinG;
-    }
-    active_tracks_posinG[feat.second->_featid] = p_FinG;
-  }
-
-  // Calibration of the first camera (cam0)
-  std::shared_ptr<Vec> distortion = state->_cam_intrinsics.at(0);
-  std::shared_ptr<PoseJPL> calibration = state->_calib_IMUtoCAM.at(0);
-  Eigen::Matrix<double, 3, 3> R_ItoC = calibration->Rot();
-  Eigen::Matrix<double, 3, 1> p_IinC = calibration->pos();
-
-  // Get current IMU clone state
-  std::shared_ptr<PoseJPL> clone_Ii = state->_clones_IMU.at(active_tracks_time);
-  Eigen::Matrix3d R_GtoIi = clone_Ii->Rot();
-  Eigen::Vector3d p_IiinG = clone_Ii->pos();
-
-  // 4. Next we can update our variable with the global position
-  //    We also will project the features into the current frame
-  for (const auto &feat : active_tracks_posinG) {
-
-    // Project the current feature into the current frame of reference
-    Eigen::Vector3d p_FinIi = R_GtoIi * (feat.second - p_IiinG);
-    Eigen::Vector3d p_FinCi = R_ItoC * p_FinIi + p_IinC;
-    double depth = p_FinCi(2);
-    Eigen::Vector2d uv_norm, uv_dist;
-    uv_norm << p_FinCi(0) / depth, p_FinCi(1) / depth;
-    uv_dist = state->_cam_intrinsics_cameras.at(0)->distort_d(uv_norm);
-
-    // Skip if not valid (i.e. negative depth, or outside of image)
-    if (depth < 0.1) {
-      continue;
-    }
-
-    // Skip if not valid (i.e. negative depth, or outside of image)
-    int width = state->_cam_intrinsics_cameras.at(0)->w();
-    int height = state->_cam_intrinsics_cameras.at(0)->h();
-    if (uv_dist(0) < 0 || (int)uv_dist(0) >= width || uv_dist(1) < 0 || (int)uv_dist(1) >= height) {
-      // PRINT_DEBUG("feat %zu -> depth = %.2f | u_d = %.2f | v_d = %.2f\n",(*it2)->featid,depth,uv_dist(0),uv_dist(1));
-      continue;
-    }
-
-    // Finally construct the uv and depth
-    Eigen::Vector3d uvd;
-    uvd << uv_dist, depth;
-    active_tracks_uvd.insert({feat.first, uvd});
-  }
-  retri_rT5 = boost::posix_time::microsec_clock::local_time();
-
-  // Timing information
-  // PRINT_DEBUG(CYAN "[RETRI-TIME]: %.4f seconds for cleaning\n" RESET, (retri_rT2-retri_rT1).total_microseconds() * 1e-6);
-  // PRINT_DEBUG(CYAN "[RETRI-TIME]: %.4f seconds for triangulate setup\n" RESET, (retri_rT3-retri_rT2).total_microseconds() * 1e-6);
-  // PRINT_DEBUG(CYAN "[RETRI-TIME]: %.4f seconds for triangulation\n" RESET, (retri_rT4-retri_rT3).total_microseconds() * 1e-6);
-  // PRINT_DEBUG(CYAN "[RETRI-TIME]: %.4f seconds for re-projection\n" RESET, (retri_rT5-retri_rT4).total_microseconds() * 1e-6);
-  // PRINT_DEBUG(CYAN "[RETRI-TIME]: %.4f seconds total\n" RESET, (retri_rT5-retri_rT1).total_microseconds() * 1e-6);
 }
