+++ ov_msckf/src/update/UpdaterSLAM.cpp
@@ -21,9 +21,42 @@
 
 #include "UpdaterSLAM.h"
 
+#include "UpdaterHelper.h"
+
+#include "feat/Feature.h"
+#include "feat/FeatureInitializer.h"
+#include "state/State.h"
+#include "state/StateHelper.h"
+#include "types/Landmark.h"
+#include "types/LandmarkRepresentation.h"
+#include "utils/colors.h"
+#include "utils/print.h"
+#include "utils/quat_ops.h"
+
+#include <boost/date_time/posix_time/posix_time.hpp>
+#include <boost/math/distributions/chi_squared.hpp>
+
 using namespace ov_core;
 using namespace ov_type;
 using namespace ov_msckf;
+
+UpdaterSLAM::UpdaterSLAM(UpdaterOptions &options_slam, UpdaterOptions &options_aruco, ov_core::FeatureInitializerOptions &feat_init_options)
+    : _options_slam(options_slam), _options_aruco(options_aruco) {
+
+  // Save our raw pixel noise squared
+  _options_slam.sigma_pix_sq = std::pow(_options_slam.sigma_pix, 2);
+  _options_aruco.sigma_pix_sq = std::pow(_options_aruco.sigma_pix, 2);
+
+  // Save our feature initializer
+  initializer_feat = std::shared_ptr<ov_core::FeatureInitializer>(new ov_core::FeatureInitializer(feat_init_options));
+
+  // Initialize the chi squared test table with confidence level 0.95
+  // https://github.com/KumarRobotics/msckf_vio/blob/050c50defa5a7fd9a04c1eed5687b405f02919b5/src/msckf_vio.cpp#L215-L221
+  for (int i = 1; i < 500; i++) {
+    boost::math::chi_squared chi_squared_dist(i);
+    chi_squared_table[i] = boost::math::quantile(chi_squared_dist, 0.95);
+  }
+}
 
 void UpdaterSLAM::delayed_init(std::shared_ptr<State> state, std::vector<std::shared_ptr<Feature>> &feature_vec) {
 
@@ -91,15 +124,15 @@ void UpdaterSLAM::delayed_init(std::shared_ptr<State> state, std::vector<std::sh
     // Triangulate the feature and remove if it fails
     bool success_tri = true;
     if (initializer_feat->config().triangulate_1d) {
-      success_tri = initializer_feat->single_triangulation_1d(it1->get(), clones_cam);
+      success_tri = initializer_feat->single_triangulation_1d(*it1, clones_cam);
     } else {
-      success_tri = initializer_feat->single_triangulation(it1->get(), clones_cam);
+      success_tri = initializer_feat->single_triangulation(*it1, clones_cam);
     }
 
     // Gauss-newton refine the feature
     bool success_refine = true;
     if (initializer_feat->config().refine_features) {
-      success_refine = initializer_feat->single_gaussnewton(it1->get(), clones_cam);
+      success_refine = initializer_feat->single_gaussnewton(*it1, clones_cam);
     }
 
     // Remove the feature if not a success
